Rename 'inifile' module to 'config'
