+++ flit/__init__.py
@@ -9,7 +9,7 @@ import sys
 from typing import Optional
 
 from flit_core import common
-from .inifile import ConfigError
+from .config import ConfigError
 from .log import enable_colourful_output
 
 __version__ = '2.3.0'
