+++ flit_core/flit_core/wheel.py
@@ -79,9 +79,9 @@ class WheelBuilder:
     @classmethod
     def from_ini_path(cls, ini_path, target_fp):
         # Local import so bootstrapping doesn't try to load pytoml
-        from . import inifile
+        from .config import read_flit_config
         directory = osp.dirname(ini_path)
-        ini_info = inifile.read_flit_config(ini_path)
+        ini_info = read_flit_config(ini_path)
         entrypoints = ini_info.entrypoints
         module = common.Module(ini_info.module, directory)
         metadata = common.make_metadata(module, ini_info)
